Implement all commands classes and reorganized QueryServlet and Main classes structure

+++ Refactoring/refactoring/src/main/java/ru/golchanskiy/sd/refactoring/html/formatting/ProductHtmlFormatting.java
@@ -17,7 +17,7 @@ public class ProductHtmlFormatting {
         return wrapBody("<h1>" + message + ": </h1>\n" + productToLocalHtml(product));
     }
 
-    public String valueToHtml(Integer value, String message) {
+    public String valueToHtml(Long value, String message) {
         return wrapBody(message + ": \n" + (value != null ? value : ""));
     }
